+++ rules/styled-components/index.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+  'rules': {
+    'block-closing-brace-newline-before': null,
+    'block-closing-brace-space-before': null,
+    'block-opening-brace-newline-after': null,
+    'block-opening-brace-newline-before': null,
+    'function-name-case': null,
+    'no-empty-source': null,
+    'no-missing-end-of-source-newline': null,
+    'value-keyword-case': null,
+  },
+};